# web/src/api/openapi.json
{
  "components": {
    "schemas": {
      "ApprovalOut": {
        "properties": {
          "candidate_id": {
            "title": "Candidate Id",
            "type": "string"
          },
          "canonical_fact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Fact Id"
          },
          "created": {
            "title": "Created",
            "type": "boolean"
          },
          "decided_at": {
            "format": "date-time",
            "title": "Decided At",
            "type": "string"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "feedback_delta_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Delta Kind"
          },
          "field_path": {
            "title": "Field Path",
            "type": "string"
          },
          "final_value": {
            "title": "Final Value"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "reviewer": {
            "title": "Reviewer",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "candidate_id",
          "field_path",
          "decision",
          "final_value",
          "reviewer",
          "note",
          "decided_at",
          "status",
          "canonical_fact_id",
          "feedback_delta_kind",
          "created"
        ],
        "title": "ApprovalOut",
        "type": "object"
      },
      "ApprovalRequest": {
        "properties": {
          "candidate_id": {
            "title": "Candidate Id",
            "type": "string"
          },
          "decision": {
            "enum": [
              "approved",
              "edited",
              "not_in_document",
              "rejected",
              "flagged"
            ],
            "title": "Decision",
            "type": "string"
          },
          "evidence": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ReviewerEvidence"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence"
          },
          "final_value": {
            "title": "Final Value"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "previous_approval_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Approval Id"
          }
        },
        "required": [
          "candidate_id",
          "decision"
        ],
        "title": "ApprovalRequest",
        "type": "object"
      },
      "ApprovalView": {
        "properties": {
          "candidate_id": {
            "title": "Candidate Id",
            "type": "string"
          },
          "decided_at": {
            "format": "date-time",
            "title": "Decided At",
            "type": "string"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "final_value": {
            "title": "Final Value"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "reviewer": {
            "title": "Reviewer",
            "type": "string"
          }
        },
        "required": [
          "id",
          "candidate_id",
          "decision",
          "final_value",
          "reviewer",
          "note",
          "decided_at"
        ],
        "title": "ApprovalView",
        "type": "object"
      },
...
      "CandidateView": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceView"
            },
            "title": "Evidence",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "prompt_name": {
            "title": "Prompt Name",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "validation": {
            "items": {
              "$ref": "#/components/schemas/ValidationView"
            },
            "title": "Validation",
            "type": "array"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "id",
          "document_id",
          "value",
          "confidence",
          "rationale",
          "status",
          "prompt_name",
          "prompt_version",
          "evidence",
          "validation"
        ],
        "title": "CandidateView",
        "type": "object"
      },
      "CanonicalFactView": {
        "properties": {
          "approval_id": {
            "title": "Approval Id",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "effective_at": {
            "format": "date-time",
            "title": "Effective At",
            "type": "string"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Evidence",
            "type": "array"
          },
          "field_path": {
            "title": "Field Path",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "object_id": {
            "title": "Object Id",
            "type": "string"
          },
          "object_type": {
            "title": "Object Type",
            "type": "string"
          },
          "object_version": {
            "title": "Object Version",
            "type": "integer"
          },
          "value": {
            "title": "Value"
          },
          "value_type": {
            "title": "Value Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "object_type",
          "object_id",
          "object_version",
          "field_path",
          "value",
          "value_type",
          "approval_id",
          "evidence",
          "effective_at",
          "created_by"
        ],
        "title": "CanonicalFactView",
        "type": "object"
      },
...
      "EvidenceView": {
        "properties": {
          "bbox": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox"
          },
          "char_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Char End"
          },
          "char_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Char Start"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "match_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Method"
          },
          "match_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Score"
          },
          "page_no": {
            "title": "Page No",
            "type": "integer"
          },
          "quote": {
            "title": "Quote",
            "type": "string"
          },
          "resolution": {
            "title": "Resolution",
            "type": "string"
          }
        },
        "required": [
          "id",
          "document_id",
          "page_no",
          "bbox",
          "char_start",
          "char_end",
          "quote",
          "resolution",
          "match_score",
          "match_method"
        ],
        "title": "EvidenceView",
        "type": "object"
      },
...
      "FieldState": {
        "properties": {
          "alternative_candidates": {
            "title": "Alternative Candidates",
            "type": "integer"
          },
          "approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalView"
              },
              {
                "type": "null"
              }
            ]
          },
          "candidate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateView"
              },
              {
                "type": "null"
              }
            ]
          },
          "enum_values": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enum Values"
          },
          "field_path": {
            "title": "Field Path",
            "type": "string"
          },
          "group": {
            "title": "Group",
            "type": "string"
          },
          "help_text": {
            "title": "Help Text",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "required": {
            "title": "Required",
            "type": "boolean"
          },
          "review_order": {
            "title": "Review Order",
            "type": "integer"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "value_type": {
            "title": "Value Type",
            "type": "string"
          }
        },
        "required": [
          "field_path",
          "label",
          "group",
          "value_type",
          "unit",
          "required",
          "help_text",
          "enum_values",
          "review_order",
          "candidate",
          "alternative_candidates",
          "approval"
        ],
        "title": "FieldState",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewState": {
        "properties": {
          "decided": {
            "title": "Decided",
            "type": "integer"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldState"
            },
            "title": "Fields",
            "type": "array"
          },
          "object_id": {
            "title": "Object Id",
            "type": "string"
          },
          "object_type": {
            "title": "Object Type",
            "type": "string"
          },
          "object_version": {
            "title": "Object Version",
            "type": "integer"
          },
          "required_undecided": {
            "title": "Required Undecided",
            "type": "integer"
          },
          "run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunView"
              },
              {
                "type": "null"
              }
            ]
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunView"
            },
            "title": "Runs",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "object_type",
          "object_id",
          "object_version",
          "run",
          "runs",
          "fields",
          "decided",
          "total",
          "required_undecided"
        ],
        "title": "ReviewState",
        "type": "object"
      },
...
      "ReviewerEvidence": {
        "description": "Evidence the reviewer gives for an edited value: the page and the text on it.",
        "properties": {
          "page_no": {
            "title": "Page No",
            "type": "integer"
          },
          "quote": {
            "title": "Quote",
            "type": "string"
          }
        },
        "required": [
          "page_no",
          "quote"
        ],
        "title": "ReviewerEvidence",
        "type": "object"
      },
      "RunView": {
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "groups": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Groups"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "schema_name": {
            "title": "Schema Name",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "document_id",
          "schema_name",
          "schema_version",
          "prompt_version",
          "groups",
          "model",
          "status"
        ],
        "title": "RunView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationView": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          },
          "rule_name": {
            "title": "Rule Name",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "rule_name",
          "passed",
          "message",
          "severity"
        ],
        "title": "ValidationView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Tender Intelligence Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/approvals": {
      "post": {
        "description": "Record a reviewer's decision. This is the only route that produces a canonical fact.\nAn identical repeat returns the existing approval with 200 and writes nothing. When\nthe request names `previous_approval_id` (null: the field was undecided) and the field\nhas been decided again since, the write is refused with 409. With a review token, only\ncandidates of the token's tender can be decided.",
        "operationId": "post_approval_api_v1_approvals_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-reviewer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Reviewer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Approval",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/canonical": {
      "get": {
        "description": "Current canonical facts of an object: approved by a human, with their evidence.",
        "operationId": "get_canonical_api_v1_canonical_get",
        "parameters": [
          {
            "in": "query",
            "name": "object_type",
            "required": true,
            "schema": {
              "title": "Object Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CanonicalFactView"
                  },
                  "title": "Response Get Canonical Api V1 Canonical Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Canonical",
        "tags": [
          "review"
        ]
      }
    },
...
    "/api/v1/review-state": {
      "get": {
        "description": "Every field of the object with its best candidate, evidence, validation and approval.",
        "operationId": "get_review_state_api_v1_review_state_get",
        "parameters": [
          {
            "in": "query",
            "name": "object_type",
            "required": true,
            "schema": {
              "title": "Object Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Review State",
        "tags": [
          "review"
        ]
      }
    },
...
  }
}